fix: carName파싱의 관심사 이전
